Drawer 3 done
Fully finished the drawer 3 prefab and added it to ownPrefab folder. Since the models are rotated, colliders on the models didn't quite work because they can't be rotated. So, I added collider children to the models. This also meant that I needed the door colliders to be triggers since the physics would be handled by the collider children. Commented out trigger skip for player interact (might have unforeseen consequences).