Fixed crouching to standing glitch
The imprecision from the player controller collider caused CanStandUp() to return false on precise calculations (i.e., trying to stand next to a table). Reduced the radius to allow for more margin of error. This causes the effect of slightly teleporting/snapping if the player is slightly inside of an object when standing up, but this is not very noticeable and less annoying than not being able to stand up.