bathroomSink
Fixed the bathroomSink prefab to match desk2. Also gave the body model an 'interactable' layer to fix a glitch where the player could not stand up near the bathroomSink. I don't think the Interactable layer does much so there should not be any unintended consequences to this, but since a layermask removes the Interactable layer from checking if a player can stand up, this is a good enough fix.